chore: remove getAssignedCards is used in my-expenses.page.spec and typo fixes (#3574)
