+++ src/app/core/mock-data/auth-response.data.ts
@@ -1,6 +1,7 @@
 import deepFreeze from 'deep-freeze-strict';
 
 import { AuthResponse } from '../models/auth-response.model';
+import { SamlResponse } from '../models/saml-response.model';
 
 export const apiAuthResponseRes: AuthResponse = deepFreeze({
   refresh_token: null,
@@ -9,4 +10,68 @@ export const apiAuthResponseRes: AuthResponse = deepFreeze({
   signup: null,
   redirect_url: null,
   org_id: null,
+});
+
+export const apiAuthRes: AuthResponse = deepFreeze({
+  cluster_domain: 'https://staging.fyle.tech',
+  access_token: null,
+  signup: false,
+  org_id: 'orDjkSfq43i2',
+  redirect_url: 'https://staging1.fyle.tech/app/accounts/#/verify?org_id=orDjkSfq43i2',
+  refresh_token:
+    'eyJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2Nzg3NzU4MzgsImlzcyI6IkZ5bGVBcHAiLCJvcmdfdXNlcl9pZCI6Ilwib3VIaW9KWVJGMmNyXCIiLCJjbHVzdGVyX2RvbWFpbiI6IlwiaHR0cHM6Ly9zdGFnaW5nLmZ5bGUudGVjaFwiIiwiZXhwIjoxOTk0MTM1ODM4fQ.FswsVpGpAt1sKI9TZ2Jxc_CqnV1K7iijoobmioH7X0o',
+});
+
+export const apiLogoutRes: AuthResponse = deepFreeze({
+  access_token:
+    'eyJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2Nzg3NzU4MzgsImlzcyI6IkZ5bGVBcHAiLCJ1c2VyX2lkIjoidXN1UVB6YVQwdzlDIiwib3JnX3VzZXJfaWQiOiJvdUhpb0pZUkYyY3IiLCJvcmdfaWQiOiJvckRqa1NmcTQzaTIiLCJyb2xlcyI6IltcIkZZTEVSXCJdIiwic2NvcGVzIjoiW10iLCJhbGxvd2VkX0NJRFJzIjoiW10iLCJ2ZXJzaW9uIjoiMyIsImNsdXN0ZXJfZG9tYWluIjoiXCJodHRwczovL3N0YWdpbmcuZnlsZS50ZWNoXCIiLCJleHAiOjE2Nzg3Nzk0Mzh9.RLRU6te7V8WkZ7AFwmSB865NDfRHUunL-VgLA4A5R44',
+  signup: null,
+  org_id: null,
+  redirect_url: null,
+  refresh_token: null,
+});
+
+export const authResData1: AuthResponse = deepFreeze({
+  cluster_domain: 'https://staging.fyle.tech',
+  access_token: null,
+  signup: false,
+  org_id: 'orNVthTo2Zyo',
+  redirect_url: 'https://staging1.fyle.tech/app/accounts/#/verify?org_id=orNVthTo2Zyo',
+  refresh_token:
+    'eyJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2Nzk5MDg3ODAsImlzcyI6IkZ5bGVBcHAiLCJvcmdfdXNlcl9pZCI6Ilwib3VYOGR3c2JMQ0x2XCIiLCJjbHVzdGVyX2RvbWFpbiI6IlwiaHR0cHM6Ly9zdGFnaW5nLmZ5bGUudGVjaFwiIiwiZXhwIjoxOTk1MjY4NzgwfQ.Ok_w8nncS6ca3_mK2PctSzMD0FluIw-PtDpn1mWn8_s',
+});
+
+export const authResData2 = deepFreeze({
+  cluster_domain: 'https://staging.fyle.tech',
+  accessToken:
+    'eyJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2NzgzNDk1NDksImlzcyI6IkZ5bGVBcHAiLCJ1c2VyX2lkIjoidXN2S0E0WDhVZ2NyIiwib3JnX3VzZXJfaWQiOiJvdVg4ZHdzYkxDTHYiLCJvcmdfaWQiOiJvck5WdGhUbzJaeW8iLCJyb2xlcyI6IltcIkFETUlOXCIsXCJBUFBST1ZFUlwiLFwiRllMRVJcIixcIkhPUFwiLFwiSE9EXCIsXCJPV05FUlwiXSIsInNjb3BlcyI6IltdIiwiYWxsb3dlZF9DSURScyI6IltdIiwidmVyc2lvbiI6IjMiLCJjbHVzdGVyX2RvbWFpbiI6IlwiaHR0cHM6Ly9zdGFnaW5nLmZ5bGUudGVjaFwiIiwiZXhwIjoxNjc4MzUzMTQ5fQ.sOJKf_ndYvhFplZL-KOImnvGujGEReQ7SYq_kvay88w',
+  signup: false,
+  org_id: 'orDjkSfq43i2',
+  redirect_url: 'https://staging1.fyle.tech/app/accounts/#/verify?org_id=orDjkSfq43i2',
+  refresh_token:
+    'eyJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2Nzg3NzU4MzgsImlzcyI6IkZ5bGVBcHAiLCJvcmdfdXNlcl9pZCI6Ilwib3VIaW9KWVJGMmNyXCIiLCJjbHVzdGVyX2RvbWFpbiI6IlwiaHR0cHM6Ly9zdGFnaW5nLmZ5bGUudGVjaFwiIiwiZXhwIjoxOTk0MTM1ODM4fQ.FswsVpGpAt1sKI9TZ2Jxc_CqnV1K7iijoobmioH7X0o',
+});
+
+export const samlResData1: SamlResponse = deepFreeze({
+  error: false,
+  response_status_code: '200',
+  cluster_domain: 'https://staging.fyle.tech',
+  access_token:
+    'eyJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2NzgzNDk1NDksImlzcyI6IkZ5bGVBcHAiLCJ1c2VyX2lkIjoidXN2S0E0WDhVZ2NyIiwib3JnX3VzZXJfaWQiOiJvdVg4ZHdzYkxDTHYiLCJvcmdfaWQiOiJvck5WdGhUbzJaeW8iLCJyb2xlcyI6IltcIkFETUlOXCIsXCJBUFBST1ZFUlwiLFwiRllMRVJcIixcIkhPUFwiLFwiSE9EXCIsXCJPV05FUlwiXSIsInNjb3BlcyI6IltdIiwiYWxsb3dlZF9DSURScyI6IltdIiwidmVyc2lvbiI6IjMiLCJjbHVzdGVyX2RvbWFpbiI6IlwiaHR0cHM6Ly9zdGFnaW5nLmZ5bGUudGVjaFwiIiwiZXhwIjoxNjc4MzUzMTQ5fQ.sOJKf_ndYvhFplZL-KOImnvGujGEReQ7SYq_kvay88w',
+  signup: false,
+  org_id: 'orDjkSfq43i2',
+  redirect_url: 'https://staging1.fyle.tech/app/accounts/#/verify?org_id=orDjkSfq43i2',
+  refresh_token:
+    'eyJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE2Nzg3NzU4MzgsImlzcyI6IkZ5bGVBcHAiLCJvcmdfdXNlcl9pZCI6Ilwib3VIaW9KWVJGMmNyXCIiLCJjbHVzdGVyX2RvbWFpbiI6IlwiaHR0cHM6Ly9zdGFnaW5nLmZ5bGUudGVjaFwiIiwiZXhwIjoxOTk0MTM1ODM4fQ.FswsVpGpAt1sKI9TZ2Jxc_CqnV1K7iijoobmioH7X0o',
+});
+
+export const samlResData2: SamlResponse = deepFreeze({
+  error: true,
+  response_status_code: '500',
+  signup: null,
+  org_id: null,
+  redirect_url: null,
+  refresh_token: null,
+  cluster_domain: null,
+  access_token: null,
 });
